+++ fiduceo/fcdr/test/writer/templates/avhrr_test.py
@@ -104,7 +104,7 @@ class AVHRRTest(unittest.TestCase):
 
         channel = ds.coords["channel"]
         self.assertEqual((6,), channel.shape)
-        self.assertEqual(5, channel[5])
+        self.assertEqual("Ch5", channel[5])
 
     def test_get_swath_width(self):
         self.assertEqual(409, AVHRR.get_swath_width())